Add integration test for Kafka interaction.
This test will start up a Kafka server (provided the script to start it is on
the PATH), run tests against it using the client, and shut it down, cleaning
up all logs and files that were created during the test.

Because the `kafka-server-start.sh` script won't be available on all hosts
that are building this package, the integration test is disabled by default.
The `integration` flag will enable it.

Also re-organized project files.